+++ src/list.tsx
@@ -227,24 +227,28 @@ interface BasicRowContentProps extends CommonProps {
 
 interface SwitchRowContentProps extends CommonProps {
     onPress?: (() => void) | undefined;
+    trackingEvent?: TrackingEvent | ReadonlyArray<TrackingEvent>;
 
     switch: ControlProps | undefined;
 }
 
 interface CheckboxRowContentProps extends CommonProps {
     onPress?: (() => void) | undefined;
+    trackingEvent?: TrackingEvent | ReadonlyArray<TrackingEvent>;
 
     checkbox: ControlProps | undefined;
 }
 
 interface RadioRowContentProps extends CommonProps {
     onPress?: (() => void) | undefined;
+    trackingEvent?: TrackingEvent | ReadonlyArray<TrackingEvent>;
 
     radioValue: string;
 }
 
 interface IconButtonRowContentProps extends CommonProps {
     onPress?: (() => void) | undefined;
+    trackingEvent?: TrackingEvent | ReadonlyArray<TrackingEvent>;
 
     iconButton: ExclusifyUnion<IconButtonProps | ToggleIconButtonProps> | undefined;
 }
@@ -475,6 +479,7 @@ const RowContent = React.forwardRef<TouchableElement, RowContentProps>((props, r
                     dataAttributes={dataAttributes}
                     disabled={disabled}
                     onPress={props.onPress}
+                    trackingEvent={props.trackingEvent}
                     role={role}
                     className={classNames(styles.dualActionLeft, {
                         [styles.touchableBackground]: hasHoverDefault,
@@ -535,14 +540,19 @@ const RowContent = React.forwardRef<TouchableElement, RowContentProps>((props, r
     if (props.iconButton) {
         return props.onPress ? (
             <div className={styles.dualActionContainer}>
-                <div
+                <BaseTouchable
+                    dataAttributes={dataAttributes}
+                    disabled={disabled}
+                    onPress={props.onPress}
+                    trackingEvent={props.trackingEvent}
+                    role={role}
                     className={classNames(styles.dualActionLeft, {
                         [styles.touchableBackground]: hasHoverDefault,
                         [styles.touchableBackgroundInverse]: hasHoverInverse,
                     })}
                 >
                     {renderContent({type: 'basic', labelId: titleId})}
-                </div>
+                </BaseTouchable>
                 <div className={styles.dualActionDivider} />
                 <Box padding={16}>
                     <Stack space="around">